resources: move log_defs from util to resources
