+++ resources/netlog_defs.ts
