+++ util/logtools/anonymizer.js
@@ -1,7 +1,7 @@
 // TODO: is the first byte of ids always flags, such that "..000000" is always empty?
 const emptyIds = ['E0000000', '80000000'];
 
-import logDefinitions from './netlog_defs';
+import logDefinitions from '../../resources/netlog_defs';
 import FakeNameGenerator from './fake_name_generator';
 
 // notifier here is a { warn: (str) => {} } object to return errors in a more structured way.
